Updated weight input step names
